requirements.txt
# Core Framework
streamlit==1.50.0
streamlit-audiorec==0.1.3

# Audio Processing
SpeechRecognition
pyttsx3
soundfile
pocketsphinx

# Vector Database and Embeddings
sentence-transformers
faiss-cpu
chromadb

# LLM Integration
google-genai
requests

# NLP and Entity Extraction
spacy==3.7.4
nltk

# Database
# sqlite3 is built into Python

# Security and Encryption
cryptography
bcrypt

# Data Processing
pandas==2.3.3
numpy<2
python-dateutil==2.9.0.post0

# Utilities
python-dotenv
# uuid, datetime, json, base64, io, os, sys, pathlib, logging, typing are built into Python
pypdf